+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v7.3.0
+------------------------------
+*December 1, 2021*
+
+### Added
+- Support for custom links in the nav via `customNavLinks` prop.
+- Relevant knobs to component story.
+
+
 v7.2.0
 ------------------------------
 *December 1, 2021*